fix(clippy): remove absurd extreme comparisons and unused comparisons

+++ src/random_problem/check_if_binary_string_has_at_most_one_segment_of_ones.rs
@@ -12,7 +12,7 @@ impl Solution {
             left += 1;
         }
 
-        while right >= 0 {
+        while right as i32 >= 0 {
             if chars[right] == '1' {
                 break;
             }
